initial attempt at inheritance, to be improved
