+++ zcash_test_vectors/transaction.py
@@ -26,9 +26,6 @@ SAPLING_TX_VERSION = 4
 NU5_VERSION_GROUP_ID = 0x26A7270A
 NU5_TX_VERSION = 5
 
-NU6_VERSION_GROUP_ID = 0x124A69F8
-NU6_TX_VERSION = 6
-
 # Sapling note magic values, copied from src/zcash/Zcash.h
 NOTEENCRYPTION_AUTH_BYTES = 16
 ZC_NOTEPLAINTEXT_LEADING = 1
@@ -314,38 +311,6 @@ class TxOut(object):
     def __bytes__(self):
         return struct.pack('<Q', self.nValue) + bytes(self.scriptPubKey)
 
-class AssetBurnDescription(object):
-    def __init__(self, rand):
-        self.AssetBase = PallasBase(leos2ip(rand.b(32)))
-        _temp = rand.u64()
-        self.valueBurn = _temp if _temp != 0 else _temp + 1
-
-    def __bytes__(self):
-        return bytes(self.AssetBase) + struct.pack('<Q', self.valueBurn)
-
-class IssueActionDescription(object):
-    def __init__(self, rand):
-        self.assetDescSize = rand.u32() % 512 + 1
-        if self.assetDescSize > 0:
-            self.asset_desc = rand.b(self.assetDescSize)
-        self.vNotes = []
-        for _ in range(rand.u8() % 5):
-            self.vNotes.append(rand.b(596))    # TODO: VA: Do we need a separate IssueNote class?
-        self.flagsIssuance = rand.u8() & 1    # Only one bit is reserved for the finalize flag currently
-
-    def __bytes__(self):
-        ret = b''
-
-        ret += struct.pack('B', int(self.assetDescSize / 256)) + struct.pack('B',self.assetDescSize % 256)
-        ret += bytes(self.asset_desc)
-        ret += write_compact_size(len(self.vNotes))
-        if len(self.vNotes) > 0:
-            for note in self.vNotes:
-                ret += note
-        ret += struct.pack('B', self.flagsIssuance)
-
-        return ret
-
 class LegacyTransaction(object):
     def __init__(self, rand, version):
         if version == OVERWINTER_TX_VERSION:
@@ -445,7 +410,6 @@ class LegacyTransaction(object):
 
         return ret
 
-
 class TransactionV5(object):
     def __init__(self, rand, consensus_branch_id):
         # Decide which transaction parts will be generated.
@@ -523,230 +487,83 @@ class TransactionV5(object):
         # <https://github.com/zcash/zcash/blob/d8c818bfa507adb845e527f5beb38345c490b330/src/primitives/transaction.h#L969-L972>
         return len(self.vin) == 1 and bytes(self.vin[0].prevout.txid) == b'\x00'*32 and self.vin[0].prevout.n == 0xFFFFFFFF
 
+    def common_txn_field_bytes(self):
+        ret = b''
+        ret += struct.pack('<I', self.version_bytes())
+        ret += struct.pack('<I', self.nVersionGroupId)
+        ret += struct.pack('<I', self.nConsensusBranchId)
+        ret += struct.pack('<I', self.nLockTime)
+        ret += struct.pack('<I', self.nExpiryHeight)
+        return ret
+
+    def transparent_txn_field_bytes(self):
+        ret = b''
+        ret += write_compact_size(len(self.vin))
+        for x in self.vin:
+            ret += bytes(x)
+        ret += write_compact_size(len(self.vout))
+        for x in self.vout:
+            ret += bytes(x)
+        return ret
+
+    def sapling_txn_field_bytes(self):
+        ret = b''
+        hasSapling = len(self.vSpendsSapling) + len(self.vOutputsSapling) > 0
+        ret += write_compact_size(len(self.vSpendsSapling))
+        for desc in self.vSpendsSapling:
+            ret += desc.bytes_v5()
+        ret += write_compact_size(len(self.vOutputsSapling))
+        for desc in self.vOutputsSapling:
+            ret += desc.bytes_v5()
+        if hasSapling:
+            ret += struct.pack('<Q', self.valueBalanceSapling)
+        if len(self.vSpendsSapling) > 0:
+            ret += bytes(self.anchorSapling)
+            # Not explicitly gated in the protocol spec, but if the gate
+            # were inactive then these loops would be empty by definition.
+            for desc in self.vSpendsSapling: # vSpendProofsSapling
+                ret += bytes(desc.proof)
+            for desc in self.vSpendsSapling: # vSpendAuthSigsSapling
+                ret += bytes(desc.spendAuthSig)
+        for desc in self.vOutputsSapling: # vOutputProofsSapling
+            ret += bytes(desc.proof)
+        if hasSapling:
+            ret += bytes(self.bindingSigSapling)
+        return ret
+
+    def orchard_txn_field_bytes(self):
+        ret = b''
+        ret += write_compact_size(len(self.vActionsOrchard))
+        if len(self.vActionsOrchard) > 0:
+            # Not explicitly gated in the protocol spec, but if the gate
+            # were inactive then these loops would be empty by definition.
+            for desc in self.vActionsOrchard:
+                ret += bytes(desc) # Excludes spendAuthSig
+            ret += struct.pack('B', self.flagsOrchard)
+            ret += struct.pack('<Q', self.valueBalanceOrchard)
+            ret += bytes(self.anchorOrchard)
+            ret += write_compact_size(len(self.proofsOrchard))
+            ret += self.proofsOrchard
+            for desc in self.vActionsOrchard:
+                ret += bytes(desc.spendAuthSig)
+            ret += bytes(self.bindingSigOrchard)
+        return ret
+
     # TODO: Update ZIP 225 to document endianness
     def __bytes__(self):
         ret = b''
 
         # Common Transaction Fields
-        ret += struct.pack('<I', self.version_bytes())
-        ret += struct.pack('<I', self.nVersionGroupId)
-        ret += struct.pack('<I', self.nConsensusBranchId)
-        ret += struct.pack('<I', self.nLockTime)
-        ret += struct.pack('<I', self.nExpiryHeight)
+        ret += self.common_txn_field_bytes()
 
         # Transparent Transaction Fields
-        ret += write_compact_size(len(self.vin))
-        for x in self.vin:
-            ret += bytes(x)
-        ret += write_compact_size(len(self.vout))
-        for x in self.vout:
-            ret += bytes(x)
+        ret += self.transparent_txn_field_bytes()
 
         # Sapling Transaction Fields
-        hasSapling = len(self.vSpendsSapling) + len(self.vOutputsSapling) > 0
-        ret += write_compact_size(len(self.vSpendsSapling))
-        for desc in self.vSpendsSapling:
-            ret += desc.bytes_v5()
-        ret += write_compact_size(len(self.vOutputsSapling))
-        for desc in self.vOutputsSapling:
-            ret += desc.bytes_v5()
-        if hasSapling:
-            ret += struct.pack('<Q', self.valueBalanceSapling)
-        if len(self.vSpendsSapling) > 0:
-            ret += bytes(self.anchorSapling)
-            # Not explicitly gated in the protocol spec, but if the gate
-            # were inactive then these loops would be empty by definition.
-            for desc in self.vSpendsSapling: # vSpendProofsSapling
-                ret += bytes(desc.proof)
-            for desc in self.vSpendsSapling: # vSpendAuthSigsSapling
-                ret += bytes(desc.spendAuthSig)
-        for desc in self.vOutputsSapling: # vOutputProofsSapling
-            ret += bytes(desc.proof)
-        if hasSapling:
-            ret += bytes(self.bindingSigSapling)
+        ret += self.sapling_txn_field_bytes()
 
         # Orchard Transaction Fields
-        ret += write_compact_size(len(self.vActionsOrchard))
-        if len(self.vActionsOrchard) > 0:
-            # Not explicitly gated in the protocol spec, but if the gate
-            # were inactive then these loops would be empty by definition.
-            for desc in self.vActionsOrchard:
-                ret += bytes(desc) # Excludes spendAuthSig
-            ret += struct.pack('B', self.flagsOrchard)
-            ret += struct.pack('<Q', self.valueBalanceOrchard)
-            ret += bytes(self.anchorOrchard)
-            ret += write_compact_size(len(self.proofsOrchard))
-            ret += self.proofsOrchard
-            for desc in self.vActionsOrchard:
-                ret += bytes(desc.spendAuthSig)
-            ret += bytes(self.bindingSigOrchard)
-
-        return ret
-
-class TransactionV6(object):
-    def __init__(self, rand, consensus_branch_id):
-        # Decide which transaction parts will be generated.
-        flip_coins = rand.u8()
-        have_transparent_in = (flip_coins >> 0) % 2
-        have_transparent_out = (flip_coins >> 1) % 2
-        have_sapling = (flip_coins >> 2) % 2
-        have_orchard_zsa = (flip_coins >> 3) % 2
-        is_coinbase = (not have_transparent_in) and (flip_coins >> 4) % 2
-        have_burn = 0
-        if have_orchard_zsa:
-            have_burn = (flip_coins >> 5) % 2
-        have_issuance = (flip_coins >> 6) % 2
-
-        # Common Transaction Fields
-        self.nVersionGroupId = NU6_VERSION_GROUP_ID
-        self.nConsensusBranchId = consensus_branch_id
-        self.nLockTime = rand.u32()
-        self.nExpiryHeight = rand.u32() % TX_EXPIRY_HEIGHT_THRESHOLD
-
-        # Transparent Transaction Fields
-        self.vin = []
-        self.vout = []
-        if have_transparent_in:
-            for _ in range((rand.u8() % 3) + 1):
-                self.vin.append(TxIn(rand))
-        if is_coinbase:
-            self.vin.append(TxIn.from_components(
-                OutPoint.from_components(b'\x00' * 32, 0xFFFFFFFF),
-                Script.coinbase_from_height(self.nExpiryHeight),
-                0xFFFFFFFF))
-        if have_transparent_out:
-            for _ in range((rand.u8() % 3) + 1):
-                self.vout.append(TxOut(rand))
-
-        # Sapling Transaction Fields
-        self.vSpendsSapling = []
-        self.vOutputsSapling = []
-        if have_sapling:
-            self.anchorSapling = Fq(leos2ip(rand.b(32)))
-            # We use the randomness unconditionally here to avoid unnecessary test vector changes.
-            for _ in range(rand.u8() % 3):
-                spend = SpendDescription(rand, self.anchorSapling)
-                if not is_coinbase:
-                    self.vSpendsSapling.append(spend)
-            for _ in range(rand.u8() % 3):
-                self.vOutputsSapling.append(OutputDescription(rand))
-            self.valueBalanceSapling = rand.u64() % (MAX_MONEY + 1)
-            self.bindingSigSapling = RedJubjubSignature(rand)
-        else:
-            # If valueBalanceSapling is not present in the serialized transaction, then
-            # v^balanceSapling is defined to be 0.
-            self.valueBalanceSapling = 0
-
-        # Orchard-ZSA Transaction Fields
-        self.vActionsOrchard = []
-        if have_orchard_zsa:
-            for _ in range(rand.u8() % 5):
-                self.vActionsOrchard.append(OrchardZSAActionDescription(rand))
-            self.flagsOrchard = rand.u8() & 7 # Only three flag bits are currently defined.
-            if is_coinbase:
-                # set enableSpendsOrchard = 0 and enableZSAs = 0
-                self.flagsOrchard &= 2
-            self.valueBalanceOrchard = rand.u64() % (MAX_MONEY + 1)
-            self.anchorOrchard = PallasBase(leos2ip(rand.b(32)))
-            self.proofsOrchard = rand.b(rand.u8() + 32) # Proof will always contain at least one element
-            self.bindingSigOrchard = RedPallasSignature(rand)
-        else:
-            # If valueBalanceOrchard is not present in the serialized transaction, then
-            # v^balanceOrchard is defined to be 0.
-            self.valueBalanceOrchard = 0
-
-        # OrchardZSA Burn Fields
-        self.vAssetBurnOrchardZSA = []
-        if have_burn:
-            for _ in range(rand.u8() % 5):
-                self.vAssetBurnOrchardZSA.append(AssetBurnDescription(rand))
-
-        # ZSA Issuance Fields
-        self.vIssueActions = []
-        if have_issuance:
-            for _ in range(rand.u8() % 5):
-                self.vIssueActions.append(IssueActionDescription(rand))
-            self.ik = rand.b(32)
-            self.issueAuthSig = rand.b(64)
-
-        assert is_coinbase == self.is_coinbase()
-
-    def version_bytes(self):
-        return NU6_TX_VERSION | (1 << 31)
-
-    def is_coinbase(self):
-        return len(self.vin) == 1 and bytes(self.vin[0].prevout.txid) == b'\x00'*32 and self.vin[0].prevout.n == 0xFFFFFFFF
-
-    def __bytes__(self):
-        ret = b''
-
-        # Common Transaction Fields
-        ret += struct.pack('<I', self.version_bytes())
-        ret += struct.pack('<I', self.nVersionGroupId)
-        ret += struct.pack('<I', self.nConsensusBranchId)
-        ret += struct.pack('<I', self.nLockTime)
-        ret += struct.pack('<I', self.nExpiryHeight)
-
-        # Transparent Transaction Fields
-        ret += write_compact_size(len(self.vin))
-        for x in self.vin:
-            ret += bytes(x)
-        ret += write_compact_size(len(self.vout))
-        for x in self.vout:
-            ret += bytes(x)
-
-        # Sapling Transaction Fields
-        hasSapling = len(self.vSpendsSapling) + len(self.vOutputsSapling) > 0
-        ret += write_compact_size(len(self.vSpendsSapling))
-        for desc in self.vSpendsSapling:
-            ret += desc.bytes_v5()
-        ret += write_compact_size(len(self.vOutputsSapling))
-        for desc in self.vOutputsSapling:
-            ret += desc.bytes_v5()
-        if hasSapling:
-            ret += struct.pack('<Q', self.valueBalanceSapling)
-        if len(self.vSpendsSapling) > 0:
-            ret += bytes(self.anchorSapling)
-            # Not explicitly gated in the protocol spec, but if the gate
-            # were inactive then these loops would be empty by definition.
-            for desc in self.vSpendsSapling: # vSpendProofsSapling
-                ret += bytes(desc.proof)
-            for desc in self.vSpendsSapling: # vSpendAuthSigsSapling
-                ret += bytes(desc.spendAuthSig)
-        for desc in self.vOutputsSapling: # vOutputProofsSapling
-            ret += bytes(desc.proof)
-        if hasSapling:
-            ret += bytes(self.bindingSigSapling)
-
-        # OrchardZSA Transaction Fields
-        ret += write_compact_size(len(self.vActionsOrchard))
-        if len(self.vActionsOrchard) > 0:
-            # Not explicitly gated in the protocol spec, but if the gate
-            # were inactive then these loops would be empty by definition.
-            for desc in self.vActionsOrchard:
-                ret += bytes(desc) # Excludes spendAuthSig
-            ret += struct.pack('B', self.flagsOrchard)
-            ret += struct.pack('<Q', self.valueBalanceOrchard)
-            ret += bytes(self.anchorOrchard)
-            ret += write_compact_size(len(self.proofsOrchard))
-            ret += self.proofsOrchard
-            for desc in self.vActionsOrchard:
-                ret += bytes(desc.spendAuthSig)
-            ret += bytes(self.bindingSigOrchard)
-
-        # OrchardZSA Burn Fields
-        ret += write_compact_size(len(self.vAssetBurnOrchardZSA))
-        if len(self.vAssetBurnOrchardZSA) > 0:
-            for desc in self.vAssetBurnOrchardZSA:
-                ret += bytes(desc)
-
-        # ZSA Issuance Fields
-        ret += write_compact_size(len(self.vIssueActions))
-        if len(self.vIssueActions) > 0:
-            for desc in self.vIssueActions:
-                ret += bytes(desc)
-            ret += bytes(self.ik)
-            ret += bytes(self.issueAuthSig)
+        ret += self.orchard_txn_field_bytes()
 
         return ret
